backend/requirement.txt
flask
pyrebase4
flask_cors
numpy
pandas
matplotlib
scikit-learn
pypickle